requirements.txt
python-daemon
gcloud
oauth2client
